Remove url data to hook and prevent setting state on unmounted
